Extract common Datum class
