+++ JsonObjects/Common/Datum.cs
@@ -1,0 +1,13 @@
+using System.Text.Json.Serialization;
+
+namespace AlarmDotCom.JsonObjects
+{
+    public class Datum
+    {
+        [JsonPropertyName("id")]
+        public string Id { get; set; }
+
+        [JsonPropertyName("type")]
+        public string Type { get; set; }
+    }
+}
