+++ JsonObjects/Systems.cs
@@ -137,13 +137,4 @@ namespace AlarmDotCom.JsonObjects.Systems
         [JsonPropertyName("meta")]
         public Meta Meta { get; set; }
     }
-
-    public class Datum
-    {
-        [JsonPropertyName("id")]
-        public string Id { get; set; }
-
-        [JsonPropertyName("type")]
-        public string Type { get; set; }
-    }
 }
